minor tweaks

+++ contrib/brushexport/export.cpp
@@ -137,7 +137,7 @@ void ExportData::GetShaderNameFromShaderPath( const char* path, std::string& nam
 	}
 
 #ifdef _DEBUG
-	globalOutputStream() << "Last: " << (const unsigned int) last_slash << " length: " << (const unsigned int)tmp.length() << "Name: " << name.c_str() << '\n';
+	globalOutputStream() << "Last: " << last_slash << " length: " << tmp.length() << "Name: " << name.c_str() << '\n';
 #endif
 }
